revising installer's support for rerunning regarding downloads and DB installation

+++ install.sh
@@ -55,7 +55,7 @@ fi
 cd "$BASEDIR"
 
 # download required files into download folder
-if [ ! -d downloads ]
+if [ ! -d downloads -o ! -f downloads/zend.tar.gz -o ! -f downloads/jpgraph.tar.gz ]
 then
   "$SCRIPT_PATH/install-download-files.sh" "$BASEDIR/downloads"
 fi
@@ -192,7 +192,7 @@ fi
 
 echo "Next you'll be now prompted to enter the root password of your MySQL server"
 $MYSQL <<LimitString
-CREATE DATABASE $DBNAME DEFAULT CHARACTER SET = UTF8 DEFAULT COLLATE = UTF8_GENERAL_CI;
+CREATE DATABASE IF NOT EXISTS $DBNAME DEFAULT CHARACTER SET = UTF8 DEFAULT COLLATE = UTF8_GENERAL_CI;
 GRANT ALL PRIVILEGES ON $DBNAME.* TO '$ADMIN'@'$MYSQLHOST' IDENTIFIED BY '$ADMIN_PASSWORD';
 GRANT SELECT,INSERT,UPDATE,DELETE ON $DBNAME.* TO '$WEBAPP_USER'@'$MYSQLHOST' IDENTIFIED BY '$WEBAPP_USER_PASSWORD';
 FLUSH PRIVILEGES;
@@ -213,18 +213,20 @@ sed -i -e "s!@db.user.name@!'$WEBAPP_USER_ESC'!" \
 
 # create createdb.sh and set database related parameters
 cd "$BASEDIR/opus4/db"
-cp createdb.sh.template createdb.sh
-if [ localhost != "$MYSQLHOST" ]; then
-  sed -i -e "s!^# host=localhost!host='$MYSQLHOST_ESC'!" createdb.sh
-fi
-if [ 3306 != "$MYSQLPORT" ]; then
-  sed -i -e "s!^# port=3306!port='$MYSQLPORT_ESC'!" createdb.sh
-fi
-sed -i -e "s!@db.admin.name@!'$ADMIN_ESC'!" \
-       -e "s!@db.admin.password@!'$ADMIN_PASSWORD_ESC'!" \
-       -e "s!@db.name@!'$DBNAME_ESC'!" createdb.sh
+if [ ! -e createdb.sh ]; then
+  cp createdb.sh.template createdb.sh
+  if [ localhost != "$MYSQLHOST" ]; then
+    sed -i -e "s!^# host=localhost!host='$MYSQLHOST_ESC'!" createdb.sh
+  fi
+  if [ 3306 != "$MYSQLPORT" ]; then
+    sed -i -e "s!^# port=3306!port='$MYSQLPORT_ESC'!" createdb.sh
+  fi
+  sed -i -e "s!@db.admin.name@!'$ADMIN_ESC'!" \
+         -e "s!@db.admin.password@!'$ADMIN_PASSWORD_ESC'!" \
+         -e "s!@db.name@!'$DBNAME_ESC'!" createdb.sh
 
-bash createdb.sh
+  bash createdb.sh || rm createdb.sh
+fi
 
 # install and configure Solr search server
 cd "$BASEDIR"
